search: factor out tests in search_results_tests.go
